+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Format based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+
+- robots.txt
+
 ## v0.1.6 - 2026-08-11
 
 ### Added